More Linux debugging
-Minor input fixes.
-Enable Vulkan validation layer for LINUX_DEBUG.
-The frame is finally rendered now but resizing has issues. Also there are black objects.

+++ Runtime/Classes/Scene.cpp
@@ -463,7 +463,7 @@ void UHScene::UpdateCamera()
 
 	if (Input->IsRightMouseHold())
 	{
-		int32_t MouseMoveX, MouseMoveY;
+		double MouseMoveX, MouseMoveY;
 		Input->GetMouseMovement(MouseMoveX, MouseMoveY);
 
 		float X = static_cast<float>(MouseMoveX) * MouseRotSpeed;
